feat: change builtin type to be u32 by it cause segfault sometimes, still investigating

+++ crates/cubecl-cpu/src/compiler/builtin.rs
@@ -2,7 +2,7 @@ use cubecl_core::CubeDim;
 
 #[derive(Default, Debug, Clone)]
 pub struct Builtin {
-    pub dims: [u64; 9],
+    pub dims: [u32; 9],
     //[
     //  cube_dim_x
     //  cube_dim_y
@@ -18,9 +18,9 @@ pub struct Builtin {
 
 impl Builtin {
     pub fn set_cube_dim(&mut self, cube_dim: CubeDim) {
-        self.dims[0] = cube_dim.x as u64;
-        self.dims[1] = cube_dim.y as u64;
-        self.dims[2] = cube_dim.z as u64;
+        self.dims[0] = cube_dim.x;
+        self.dims[1] = cube_dim.y;
+        self.dims[2] = cube_dim.z;
     }
     pub fn get_cube_dim(&self) -> (u32, u32, u32) {
         (
@@ -30,16 +30,14 @@ impl Builtin {
         )
     }
     pub fn set_cube_count(&mut self, cube_count: [u32; 3]) {
-        self.dims[3] = cube_count[0] as u64;
-        self.dims[4] = cube_count[1] as u64;
-        self.dims[5] = cube_count[2] as u64;
+        self.dims[3] = cube_count[0];
+        self.dims[4] = cube_count[1];
+        self.dims[5] = cube_count[2];
     }
     pub fn set_unit_pos(&mut self, unit_pos: [u32; 3]) {
-        // TODO fix alignment issue and pass dims as u32 instead of u64
-        // self.dims[6..9].copy_from_slice(unit_pos);
-        self.dims[6] = unit_pos[0] as u64;
-        self.dims[7] = unit_pos[1] as u64;
-        self.dims[8] = unit_pos[2] as u64;
+        self.dims[6] = unit_pos[0];
+        self.dims[7] = unit_pos[1];
+        self.dims[8] = unit_pos[2];
     }
     pub const fn len(&self) -> usize {
         self.dims.len()
